renamed interfaces' bycurrency to by_currency
